# src/swagger/swagger_temp.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backend API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3001/backend",
  "basePath": "/api/v1/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/company": {
      "get": {
        "tags": [
          "Company"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "post": {
        "tags": [
          "Company"
        ],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "body",
            "description": "new object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": ""
                },
                "acronym": {
                  "type": "string",
                  "example": ""
                },
                "active": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          },
          {
            "name": "login",
            "description": "login",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "password",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    },
    "/company/{id}": {
      "get": {
        "tags": [
          "Company"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "delete": {
        "tags": [
          "Company"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "patch": {
        "tags": [
          "Company"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "body",
            "description": "update object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": ""
                },
                "acronym": {
                  "type": "string",
                  "example": ""
                },
                "active": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          },
          {
            "name": "login",
            "description": "login",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "password",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "body",
            "description": "new object",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": ""
                },
                "password": {
                  "type": "string",
                  "example": ""
                },
                "active": {
                  "type": "number",
                  "example": 0
                },
                "company_id": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          {
            "name": "login",
            "description": "login",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "password",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "body",
            "description": "update object",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": ""
                },
                "password": {
                  "type": "string",
                  "example": ""
                },
                "active": {
                  "type": "number",
                  "example": 0
                },
                "company_id": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          {
            "name": "login",
            "description": "login",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "password",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    },
    "/unit": {
      "get": {
        "tags": [
          "Unit"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "post": {
        "tags": [
          "Unit"
        ],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "body",
            "description": "new object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": ""
                },
                "company_id": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          {
            "name": "login",
            "description": "login",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "password",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    },
    "/unit/{id}": {
      "get": {
        "tags": [
          "Unit"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "delete": {
        "tags": [
          "Unit"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "patch": {
        "tags": [
          "Unit"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "body",
            "description": "update object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": ""
                },
                "company_id": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          {
            "name": "login",
            "description": "login",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "password",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    },
    "/asset": {
      "get": {
        "tags": [
          "Asset"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "post": {
        "tags": [
          "Asset"
        ],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "body",
            "description": "new object",
            "schema": {
              "type": "object",
              "properties": {
                "unit_id": {
                  "type": "string",
                  "example": ""
                },
                "image": {
                  "type": "string",
                  "example": ""
                },
                "name": {
                  "type": "string",
                  "example": ""
                },
                "description": {
                  "type": "string",
                  "example": ""
                },
                "model": {
                  "type": "string",
                  "example": ""
                },
                "owner": {
                  "type": "string",
                  "example": ""
                },
                "status": {
                  "type": "string",
                  "example": ""
                },
                "health_level": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          {
            "name": "login",
            "description": "login",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "password",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    },
    "/asset/{id}": {
      "get": {
        "tags": [
          "Asset"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "delete": {
        "tags": [
          "Asset"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      },
      "patch": {
        "tags": [
          "Asset"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "body",
            "description": "update object",
            "schema": {
              "type": "object",
              "properties": {
                "unit_id": {
                  "type": "string",
                  "example": ""
                },
                "image": {
                  "type": "string",
                  "example": ""
                },
                "name": {
                  "type": "string",
                  "example": ""
                },
                "description": {
                  "type": "string",
                  "example": ""
                },
                "model": {
                  "type": "string",
                  "example": ""
                },
                "owner": {
                  "type": "string",
                  "example": ""
                },
                "status": {
                  "type": "string",
                  "example": ""
                },
                "health_level": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          {
            "name": "login",
            "description": "login",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "password",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    },
    "/upload": {
      "post": {
        "tags": [
          "Misc"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "imgID": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Misc"
        ],
        "description": "",
        "parameters": [
          {
            "name": "login",
            "description": "login",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "password",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error on server"
          }
        }
      }
    }
  }
}